Adding the document, and moved some more files around to make the xrefs in the doc make sense
