+++ mcsema/Arch/X86/Semantics/CMOV.h
@@ -27,6 +27,7 @@ ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#include "InstructionDispatch.h"
+
+class DispatchMap;
 
 void CMOV_populateDispatchMap(DispatchMap &m);
