+++ mcsema/Arch/X86/Semantics/INCDECNEG.cpp
@@ -46,10 +46,11 @@
 #include "mcsema/Arch/Dispatch.h"
 #include "mcsema/Arch/Register.h"
 
-#include "InstructionDispatch.h"
 #include "mcsema/Arch/X86/Util.h"
 #include "mcsema/Arch/X86/Semantics/flagops.h"
 #include "mcsema/Arch/X86/Semantics/INCDECNEG.h"
+
+#include "mcsema/BC/Util.h"
 
 #define NASSERT(cond) TASSERT(cond, "")
