+++ mcsema/Arch/X86/Semantics/Jcc.cpp
@@ -49,9 +49,10 @@
 #include "mcsema/Arch/Dispatch.h"
 #include "mcsema/Arch/Register.h"
 
-#include "InstructionDispatch.h"
 #include "mcsema/Arch/X86/Util.h"
 #include "mcsema/Arch/X86/Semantics/Jcc.h"
+
+#include "mcsema/BC/Util.h"
 
 #define NASSERT(cond) TASSERT(cond, "")
