+++ mcsema/Arch/X86/Semantics/MOV.cpp
@@ -49,8 +49,10 @@
 #include "mcsema/Arch/X86/Util.h"
 #include "mcsema/Arch/X86/Semantics/MOV.h"
 
-#include "mcsema/cfgToLLVM/Externals.h"
-#include "mcsema/cfgToLLVM/InstructionDispatch.h"
+#include "mcsema/BC/Util.h"
+
+#include "mcsema/CFG/Externals.h"
+
 #include "mcsema/cfgToLLVM/JumpTables.h"
 
 #define NASSERT(cond) TASSERT(cond, "")
