+++ mcsema/Arch/X86/Semantics/MULDIV.cpp
@@ -46,9 +46,10 @@
 #include "mcsema/Arch/Dispatch.h"
 #include "mcsema/Arch/Register.h"
 
-#include "InstructionDispatch.h"
 #include "mcsema/Arch/X86/Util.h"
 #include "mcsema/Arch/X86/Semantics/MULDIV.h"
+
+#include "mcsema/BC/Util.h"
 
 #define NASSERT(cond) TASSERT(cond, "")
