+++ mcsema/Arch/X86/Semantics/String.cpp
@@ -46,10 +46,11 @@
 #include "mcsema/Arch/Dispatch.h"
 #include "mcsema/Arch/Register.h"
 
-#include "InstructionDispatch.h"
 #include "mcsema/Arch/X86/Util.h"
 #include "mcsema/Arch/X86/Semantics/flagops.h"
 #include "mcsema/Arch/X86/Semantics/String.h"
+
+#include "mcsema/BC/Util.h"
 
 template<int width, int regWidth>
 static llvm::BasicBlock *doCmpsV(llvm::BasicBlock *pred) {
