+++ mcsema/Arch/X86/Semantics/fpu.cpp
@@ -51,7 +51,8 @@
 
 #include "mcsema/Arch/X86/Util.h"
 #include "mcsema/Arch/X86/Semantics/fpu.h"
-#include "mcsema/cfgToLLVM/InstructionDispatch.h"
+
+#include "mcsema/BC/Util.h"
 
 #ifndef M_PIl
 # define M_PIl    3.141592653589793238462643383279502884L /* pi */
