+++ mcsema/cfgToLLVM/JumpTables.cpp
@@ -46,10 +46,10 @@
 #include "mcsema/Arch/X86/Util.h"  // TODO(pag): MEM_AS_DATA_REF
 
 #include "mcsema/BC/Util.h"
+
 #include "mcsema/CFG/CFG.h"
 
 #include "mcsema/cfgToLLVM/JumpTables.h"
-#include "mcsema/cfgToLLVM/InstructionDispatch.h"
 #include "mcsema/cfgToLLVM/TransExcn.h"
 
 // convert a jump table to a data section of symbols
